+++ packages/react/src/components/CardEditor/CardEditForm/CardEditFormItems/DataSeriesFormItemModal.jsx
@@ -176,9 +176,10 @@ const DataSeriesFormItemModal = ({
 
   const isTimeBasedCard =
     type === CARD_TYPES.TIMESERIES ||
-    type === CARD_TYPES.TABLE ||
-    content?.type === BAR_CHART_TYPES.SIMPLE ||
-    content?.type === BAR_CHART_TYPES.STACKED;
+    (type === CARD_TYPES.TABLE &&
+      content?.columns?.find((column) => column.type === 'TIMESTAMP')) ||
+    (content?.type === BAR_CHART_TYPES.SIMPLE && content?.timeDataSourceId) ||
+    (content?.type === BAR_CHART_TYPES.STACKED && content?.timeDataSourceId);
 
   const handleTranslation = useCallback(
     (idToTranslate) => {
@@ -394,7 +395,7 @@ const DataSeriesFormItemModal = ({
       ) : null}
 
       <div className={`${baseClassName}--input-group`}>
-        <div className={`${baseClassName}--input-group--item`}>
+        <div className={`${baseClassName}--input-group--item-half`}>
           <TextInput
             id={`${id}_attribute-label`}
             labelText={mergedI18n.dataItemEditorDataItemCustomLabel}
@@ -411,7 +412,7 @@ const DataSeriesFormItemModal = ({
         </div>
 
         {(type === CARD_TYPES.TIMESERIES || type === CARD_TYPES.BAR) && ( // show color selector
-          <div className={`${baseClassName}--input-group--item`}>
+          <div className={`${baseClassName}--input-group--item-half`}>
             <ColorDropdown
               id={`${id}_color-dropdown`}
               label=""