+++ packages/react/src/components/Table/Table.story.jsx
@@ -2622,6 +2622,7 @@ export const WithSorting = () => (
     columns={tableColumns.map((i, idx) => ({
       ...i,
       isSortable: idx !== 1,
+      align: i.id === 'number' ? 'end' : i.id === 'string' ? 'center' : 'start',
     }))}
     data={getSortedData(tableData, 'string', 'ASC')}
     actions={actions}
@@ -2629,9 +2630,20 @@ export const WithSorting = () => (
       hasFilter: false,
       hasPagination: true,
       hasRowSelection: 'multi',
+      hasAggregations: true,
     }}
     view={{
       filters: [],
+      aggregations: {
+        label: 'Total',
+        columns: [
+          {
+            id: 'number',
+            align: 'end',
+            isSortable: true,
+          },
+        ],
+      },
       table: {
         ordering: defaultOrdering,
         sort: {